chart-analysis-pipeline/chart-analysis-ps/requirements.txt
clarifai>=12.0.0
Pillow>=10.0.0
requests>=2.31.0
